Connection to Exam 4.8 Server fixed. Failed due tue SOAP 1.2
